Add files via upload
This is an Arduino-based project that controls and monitors water levels using an ultrasonic sensor. It displays the distance on an LCD, controls a pump using a relay, and is fully simulated on the Wokwi platform. This repo includes all circuit diagrams, code, and links for simulation. Designed for automation beginners and IoT learners.